requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cssselect==1.2.0
cycler==0.11.0
discord.py==2.2.2
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-file==1.5.1
six==1.16.0
sportsreference==0.5.2
tldextract==3.4.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.8.2
